refactor(wallet-mobile): consider delegating banner
