+++ apps/wallet-mobile/src/features/Banners/common/strings.ts
