+++ apps/wallet-mobile/translations/messages/src/features/Banners/common/strings.json
@@ -1,0 +1,47 @@
+[
+  {
+    "id": "components.considerDelegatingToYoroiBanner.description",
+    "defaultMessage": "!!!Delegate to our DRep and help Cardano evolve in a way that benefits your wallet experience.",
+    "file": "src/features/Banners/common/strings.ts",
+    "start": {
+      "line": 13,
+      "column": 15,
+      "index": 327
+    },
+    "end": {
+      "line": 16,
+      "column": 3,
+      "index": 516
+    }
+  },
+  {
+    "id": "components.considerDelegatingToYoroiBanner.title",
+    "defaultMessage": "!!!Consider delegating to Yoroi",
+    "file": "src/features/Banners/common/strings.ts",
+    "start": {
+      "line": 17,
+      "column": 9,
+      "index": 527
+    },
+    "end": {
+      "line": 20,
+      "column": 3,
+      "index": 647
+    }
+  },
+  {
+    "id": "components.considerDelegatingToYoroiBanner.cta",
+    "defaultMessage": "!!!Delegate Now",
+    "file": "src/features/Banners/common/strings.ts",
+    "start": {
+      "line": 21,
+      "column": 7,
+      "index": 656
+    },
+    "end": {
+      "line": 24,
+      "column": 3,
+      "index": 758
+    }
+  }
+]
